fix: fix uid lookup on logs / metrics and relocate one-time utils to correct file
